Adding some minor structural changes.
