--- TollFeeSystem.Simulator/Simulator.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TollFeeSystem.Simulator
-{
-    internal class Simulator
-    {
-    }
-}
