--- TollFeeSystem.Simulator/TheWorld.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TollFeeSystem.Simulator
-{
-    internal class TheWorld
-    {
-
-    }
-}
