+++ TollFeeSystem.Simulator/ToDo.txt
@@ -11,3 +11,8 @@ in memory db
 
 
 är github publik?
+
+
+Enfilig och tvåfilig portaler
+
+en tvåfilig kan ta två fordon i teorin, dubbla kameror?
